renamed Checkbox to Filter
